pyproject.toml
[tool.poetry]
name = "test-21vek"
version = "0.1.0"
description = ""
authors = ["Vitali Mitsevich"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
fastapi = {extras = ["all"], version = "^0.115.0"}
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.35"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
